chore: rename helper and add JSDoc comments
