+++ packages/shared/src/utilities/react/index.ts
@@ -1,4 +1,4 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './joinClassNames';
-export * from './updateArray';
+export * from './moveItemInArray';
