+++ packages/shared/src/utilities/react/moveItemInArray.spec.ts
@@ -1,9 +1,9 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { describe, expect, it } from 'vitest';
-import { updateArray } from './updateArray';
+import { moveItemInArray } from './moveItemInArray';
 
-describe('updateArray', () => {
+describe('moveItemInArray', () => {
     it.each([
         [[], 0, 0, []],
         [[1, 2, 3], 0, 0, [1, 2, 3]],
@@ -12,6 +12,6 @@ describe('updateArray', () => {
         [[1, { a: 1 }, 3], 1, 0, [{ a: 1 }, 1, 3]],
         [[{ a: 1 }, { b: 1 }, { c: 1 }], 1, 0, [{ b: 1 }, { a: 1 }, { c: 1 }]],
     ])('move items accordingly', (items, from, to, expected) => {
-        expect(expected).toEqual(updateArray(items as number[], from, to));
+        expect(expected).toEqual(moveItemInArray(items as number[], from, to));
     });
 });
